Move sign in/up to auth library.
